requirements.txt
beautifulsoup4==4.11.1
Brotli==1.0.9
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Cython==0.29.32
dash==2.6.1
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.2
Flask-Compress==1.12
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.23.2
packaging==21.3
pandas==1.4.3
pandas-datareader==0.10.0
patsy==0.5.2
plotly==5.10.0
pmdarima==1.8.5
proxy-requests==0.5.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
soupsieve==2.3.2.post1
statsmodels==0.13.2
tenacity==8.0.1
threadpoolctl==3.1.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
